app/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-services-account-tools==1.4.0
autogluon.tabular
pydantic